+++ src/main/java/ru/dpopkov/tasktracker/Action.java
@@ -1,0 +1,10 @@
+package ru.dpopkov.tasktracker;
+
+public interface Action {
+
+    ActionType type();
+
+    String description();
+
+    void execute();
+}