+++ src/main/java/ru/dpopkov/tasktracker/ActionType.java
@@ -1,0 +1,8 @@
+package ru.dpopkov.tasktracker;
+
+public enum ActionType {
+    ADD_USER,
+    SHOW_ALL_USERS,
+    SHOW_USER,
+    EXIT_APP
+}