+++ src/main/java/ru/dpopkov/tasktracker/Main.java
@@ -1,9 +1,23 @@
 package ru.dpopkov.tasktracker;
 
+import ru.dpopkov.tasktracker.repositories.inmemory.SimpleUserRepository;
+import ru.dpopkov.tasktracker.services.UserServiceImpl;
+import ru.dpopkov.tasktracker.ui.*;
+import ru.dpopkov.tasktracker.ui.actions.Actions;
+import ru.dpopkov.tasktracker.ui.actions.ActionPack;
+
 public class Main {
 
     public static void main(String[] args) {
-        MessageProvider messageProvider = new MessageProvider();
-        System.out.println("Message: " + messageProvider.getMessage());
+        TaskTracker taskTracker = new TaskTracker(new UserServiceImpl(new SimpleUserRepository()));
+
+        UiInput input = new ConsoleUiInput();
+        UiOutput output = System.out::print;
+        ActionPack pack = new ActionPack(input, output, taskTracker);
+        Actions actions = new Actions(pack);
+
+        Menu menu = new TextualMenu(actions, output, input);
+        UiCycle uiCycle = new UiCycle(actions, menu);
+        uiCycle.run();
     }
 }