+++ src/main/java/ru/dpopkov/tasktracker/repositories/inmemory/SimpleUserRepository.java
@@ -1,6 +1,7 @@
 package ru.dpopkov.tasktracker.repositories.inmemory;
 
 import ru.dpopkov.tasktracker.model.User;
+import ru.dpopkov.tasktracker.repositories.UserRepository;
 
-public class SimpleUserRepository extends SimpleBaseRepository<User> {
+public class SimpleUserRepository extends SimpleBaseRepository<User> implements UserRepository {
 }