+++ src/main/java/ru/dpopkov/tasktracker/services/UserService.java
@@ -1,0 +1,6 @@
+package ru.dpopkov.tasktracker.services;
+
+import ru.dpopkov.tasktracker.model.User;
+
+public interface UserService extends BaseService<User> {
+}