+++ src/main/java/ru/dpopkov/tasktracker/services/UserServiceImpl.java
@@ -1,0 +1,11 @@
+package ru.dpopkov.tasktracker.services;
+
+import ru.dpopkov.tasktracker.model.User;
+import ru.dpopkov.tasktracker.repositories.UserRepository;
+
+public class UserServiceImpl extends AbstractBaseService<User> implements UserService {
+
+    public UserServiceImpl(UserRepository repository) {
+        super(repository);
+    }
+}