+++ src/main/java/ru/dpopkov/tasktracker/ui/UiInput.java
@@ -1,0 +1,8 @@
+package ru.dpopkov.tasktracker.ui;
+
+public interface UiInput {
+
+    String readString();
+
+    int readInt();
+}