+++ src/main/java/ru/dpopkov/tasktracker/ui/UiOutput.java
@@ -1,0 +1,6 @@
+package ru.dpopkov.tasktracker.ui;
+
+public interface UiOutput {
+
+    void print(String s);
+}