+++ History.md
@@ -5,3 +5,4 @@ History of implementation
 2. Add config for unit and integration tests.
 3. Add Model: User, Project, Task.
 4. Add interfaces for repositories and simple in-memory implementations.
+5. Implement UI, add service for User and implement operations of adding and viewing all users.